pyproject.toml
[project]
name = "fastapi-multitenant"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Sayanc2000", email = "sayanc20002@gmail.com" }
]
dependencies = [
    "fastapi>=0.115.0",
    "sqlalchemy>=2.0.35",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.31.0",
    "psycopg2-binary>=2.9.9",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/fastapi_multitenant"]
